+++ CHANGELOG.md
@@ -5,7 +5,8 @@
 #### Enhancements
 - Added new venues, datasets, and publishers for US Equities Securities Information Processors
 - Changed the `batch.download` method to use the API's `batch.download` endpoint for
-downloading a batch job as a ZIP archive
+  downloading a batch job as a ZIP archive
+- Added method for `batch.get_job_details` to access the full details of a batch job
 
 ## 0.78.0 - 2026-05-12